requirements.txt
discord.py==2.5.2
asyncpg==0.30.0
tortoise-orm==0.25.1
python-dotenv==1.0.1
requests>=2.31.0
matplotlib
